Перенос моделей и моделей-представления в проект Предметной области
