+++ WebStore/Common/WebStore.Domain/Entities/Orders/OrderItem.cs
@@ -1,6 +1,7 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 using WebStore.Domain.Entities.Base;
+using WebStore.Domain.Entities.Products;
 
 namespace WebStore.Domain.Entities.Orders
 {
