+++ WebStore/Common/WebStore.Domain/Entities/Products/Brand.cs
@@ -2,7 +2,7 @@
 using WebStore.Domain.Entities.Base;
 using WebStore.Domain.Entities.Base.Interfaces;
 
-namespace WebStore.Domain.Entities
+namespace WebStore.Domain.Entities.Products
 {
     /// <summary>
     /// Бренд
@@ -10,7 +10,7 @@ namespace WebStore.Domain.Entities
     public class Brand : NamedEntity, IOrderedEntity
     {
         public int Order { get; set; }
-        
+
         /// <summary>
         /// Список товаров данного бренда
         /// </summary>
