+++ WebStore/Common/WebStore.Domain/Models/CartItem.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-
-namespace WebStore.Models
+﻿namespace WebStore.Domain.Models
 {
     /// <summary>
     /// Элемент корзины
