+++ WebStore/Common/WebStore.Domain/ViewModels/Identity/RegisterUserViewModel.cs
@@ -1,9 +1,8 @@
 ﻿using System.ComponentModel.DataAnnotations;
-using Microsoft.AspNetCore.Mvc;
 
-namespace WebStore.ViewModels.Identity
+namespace WebStore.Domain.ViewModels.Identity
 {
-    public class LoginViewModel
+    public class RegisterUserViewModel
     {
         [Required]
         [MaxLength(256)]
@@ -15,10 +14,10 @@ namespace WebStore.ViewModels.Identity
         [Display(Name = "Пароль")]
         public string Password { get; set; }
 
-        [Display(Name = "Запомнить меня!")]
-        public bool RememberMe { get; set; }
-
-        [HiddenInput(DisplayValue = false)]
-        public string ReturnUrl { get; set; }
+        [Required]
+        [DataType(DataType.Password)]
+        [Display(Name = "Подтвердите ввод пароля")]
+        [Compare(nameof(Password))]
+        public string ConfirmPassword { get; set; }
     }
 }
