+++ WebStore/Services/WebStore.DAL/Context/WebStoreDB.cs
@@ -1,9 +1,9 @@
 ﻿using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
-using WebStore.Domain.Entities;
 using WebStore.Domain.Entities.Blog;
 using WebStore.Domain.Entities.Identity;
 using WebStore.Domain.Entities.Orders;
+using WebStore.Domain.Entities.Products;
 
 namespace WebStore.DAL.Context
 {
