+++ WebStore/UI/WebStore/Areas/Admin/Views/_ViewImports.cshtml
@@ -1,7 +1,8 @@
-﻿@using WebStore.Models
-@using WebStore.ViewModels
-@using WebStore.ViewModels.Identity
-@using WebStore.Domain.Entities
+﻿@using WebStore.Domain.Models
+@using WebStore.Domain.ViewModels
+@using WebStore.Domain.ViewModels.Orders
+@using WebStore.Domain.ViewModels.Identity
+@using WebStore.Domain.Entities.Products
 
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @addTagHelper *, WebStore
