+++ WebStore/UI/WebStore/Components/BrandsViewComponent.cs
@@ -3,9 +3,10 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
+using WebStore.Domain.ViewModels.Product;
 using WebStore.Infrastructure.Interfaces;
 using WebStore.Services;
-using WebStore.ViewModels;
+
 
 namespace WebStore.Components
 {
