+++ WebStore/UI/WebStore/Controlers/UserProfileController.cs
@@ -3,8 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
+using WebStore.Domain.ViewModels.Orders;
 using WebStore.Infrastructure.Interfaces;
-using WebStore.ViewModels.Orders;
+
 
 namespace WebStore.Controllers
 {
