+++ WebStore/UI/WebStore/Infrastructure/Interfaces/ICartService.cs
@@ -2,7 +2,8 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using WebStore.ViewModels;
+using WebStore.Domain.ViewModels.Product;
+
 
 namespace WebStore.Infrastructure.Interfaces
 {
