+++ WebStore/UI/WebStore/Infrastructure/Interfaces/IProductData.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using WebStore.Domain.Entities;
+using WebStore.Domain.Entities.Products;
 
 namespace WebStore.Infrastructure.Interfaces
 {
