+++ WebStore/UI/WebStore/Infrastructure/Mapping/Blog/BlogCommentMapping.cs
@@ -1,5 +1,6 @@
 ﻿using WebStore.Domain.Entities.Blog;
-using WebStore.ViewModels.Blog;
+using WebStore.Domain.ViewModels.Blog;
+
 
 namespace WebStore.Infrastructure.Mapping.Blog
 {
