+++ WebStore/UI/WebStore/Infrastructure/Mapping/Blog/BlogRatingMapping.cs
@@ -3,7 +3,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using WebStore.Domain.Entities.Blog;
-using WebStore.ViewModels.Blog;
+using WebStore.Domain.ViewModels.Blog;
+
 
 namespace WebStore.Infrastructure.Mapping.Blog
 {
