+++ WebStore/UI/WebStore/Infrastructure/Mapping/ProductMapping.cs
@@ -1,7 +1,9 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using WebStore.Domain.Entities;
-using WebStore.ViewModels;
+using WebStore.Domain.Entities.Products;
+using WebStore.Domain.ViewModels.Product;
+
 
 namespace WebStore.Infrastructure.Mapping
 {
