+++ WebStore/UI/WebStore/Infrastructure/Services/InCookies/CookiesCartService.cs
@@ -4,10 +4,12 @@ using System.Linq;
 using Microsoft.AspNetCore.Http;
 using Newtonsoft.Json;
 using WebStore.Domain.Entities;
+using WebStore.Domain.Entities.Products;
+using WebStore.Domain.Models;
+using WebStore.Domain.ViewModels.Product;
 using WebStore.Infrastructure.Interfaces;
 using WebStore.Infrastructure.Mapping;
-using WebStore.Models;
-using WebStore.ViewModels;
+
 
 namespace WebStore.Infrastructure.Services.InCookies
 {
