+++ WebStore/UI/WebStore/Infrastructure/Services/InSQL/SqlProductData.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using WebStore.DAL.Context;
 using WebStore.Domain.Entities;
+using WebStore.Domain.Entities.Products;
 using WebStore.Infrastructure.Interfaces;
 
 namespace WebStore.Infrastructure.Services.InSQL
