+++ WebStore/UI/WebStore/Views/UserProfile/Orders.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<WebStore.ViewModels.Orders.UserOrderViewModel>
+﻿@model IEnumerable<UserOrderViewModel>
 @{
     Layout = "_UserProfileLayout";
     ViewData["Title"] = "Заказы пользователя";
