+++ WebStore/UI/WebStore/Views/_ViewImports.cshtml
@@ -1,7 +1,8 @@
-﻿@using WebStore.Models
-@using WebStore.ViewModels
-@using WebStore.ViewModels.Blog
-@using WebStore.ViewModels.Identity
+﻿@using WebStore.Domain.Models
+@using WebStore.Domain.ViewModels.Orders
+@using WebStore.Domain.ViewModels.Product
+@using WebStore.Domain.ViewModels.Blog
+@using WebStore.Domain.ViewModels.Identity
 @using WebStore.Domain.Entities.Identity
